refactor(emit): split ir_printer and class_es5_ir below 2000-line §19 cap
Extract large impl blocks into dedicated submodule files:

- ir_printer_class_emit.rs: ES5 class, async/generator, and ASTRef emit
  helpers (8 match arms delegated from emit_node in ir_printer.rs)
- class_es5_ir_decorators.rs: legacy + TC39 decorator lowering helpers
- class_es5_ir_auto_accessor.rs: auto-accessor, property, and parameter
  helpers

All five touched hand-authored files are now under 2000 lines:
  ir_printer.rs        2607 → 1827
  ir_printer_class_emit.rs   (new, 810)
  class_es5_ir.rs      3451 → 1787
  class_es5_ir_decorators.rs (new, 811)
  class_es5_ir_auto_accessor.rs (new, 899)

No logic changes; all 3269 pre-existing tests pass/fail identically.
