+++ crates/tsz-emitter/src/transforms/class_es5_ir.rs
@@ -73,8 +73,12 @@
 pub mod ast_to_ir;
 pub use ast_to_ir::AstToIr;
 
+#[path = "class_es5_ir_auto_accessor.rs"]
+mod auto_accessor;
 #[path = "class_es5_ir_constructor.rs"]
 mod constructor;
+#[path = "class_es5_ir_decorators.rs"]
+mod decorators;
 #[path = "class_es5_ir_helpers.rs"]
 mod helpers;
 #[path = "class_es5_ir_members.rs"]
@@ -84,8 +88,7 @@ use helpers::*;
 use crate::context::transform::TransformContext;
 use crate::transforms::async_es5_ir::AsyncES5Transformer;
 use crate::transforms::ir::{
-    IRCatchClause, IRMethodName, IRNode, IRParam, IRProperty, IRPropertyDescriptor, IRPropertyKey,
-    IRPropertyKind, IRSwitchCase,
+    IRCatchClause, IRNode, IRParam, IRProperty, IRPropertyKey, IRPropertyKind, IRSwitchCase,
 };
 use crate::transforms::ir_printer::IRPrinter;
 use crate::transforms::private_fields_es5::{
@@ -98,7 +101,7 @@ use tsz_common::common::ModuleKind;
 use tsz_parser::parser::node::{Node, NodeAccess, NodeArena};
 use tsz_parser::parser::syntax_kind_ext;
 use tsz_parser::parser::{NodeIndex, NodeList};
-use tsz_parser::syntax::transform_utils::{contains_new_target_reference, contains_this_reference};
+use tsz_parser::syntax::transform_utils::contains_this_reference;
 use tsz_scanner::SyntaxKind;
 
 struct Tc39Es5MemberDecorator {
@@ -1100,798 +1103,6 @@ impl<'a> ES5ClassTransformer<'a> {
         None
     }
 
-    /// Collect decorator `NodeIndex` list from a modifier list
-    fn collect_decorators_from_modifiers(&self, modifiers: &Option<NodeList>) -> Vec<NodeIndex> {
-        let Some(mods) = modifiers else {
-            return Vec::new();
-        };
-        mods.nodes
-            .iter()
-            .copied()
-            .filter(|&mod_idx| {
-                self.arena
-                    .get(mod_idx)
-                    .is_some_and(|n| n.kind == syntax_kind_ext::DECORATOR)
-            })
-            .collect()
-    }
-
-    /// Collect parameter decorators from a method's parameter list for ES5 emit.
-    /// Returns `Vec` of (`runtime_param_index`, `decorator_node_indices`).
-    /// Skips the `this` parameter since it's erased in JS emit.
-    fn collect_param_decorators_es5(&self, parameters: &NodeList) -> Vec<(usize, Vec<NodeIndex>)> {
-        let mut result = Vec::new();
-        let mut runtime_index = 0usize;
-        for &param_idx in &parameters.nodes {
-            let Some(param_node) = self.arena.get(param_idx) else {
-                continue;
-            };
-            let Some(param) = self.arena.get_parameter(param_node) else {
-                continue;
-            };
-
-            // Skip `this` parameter
-            if let Some(name_node) = self.arena.get(param.name) {
-                if name_node.kind == SyntaxKind::ThisKeyword as u16 {
-                    continue;
-                }
-                if name_node.kind == SyntaxKind::Identifier as u16
-                    && self
-                        .arena
-                        .get_identifier(name_node)
-                        .is_some_and(|id| id.escaped_text == "this")
-                {
-                    continue;
-                }
-            }
-
-            let decorators = self.collect_decorators_from_modifiers(&param.modifiers);
-            if !decorators.is_empty() {
-                result.push((runtime_index, decorators));
-            }
-            runtime_index += 1;
-        }
-        result
-    }
-
-    fn helper_name(&self, name: &str) -> String {
-        if self.tslib_prefix {
-            format!("{}.{name}", self.tslib_import_binding)
-        } else {
-            name.to_string()
-        }
-    }
-
-    /// Render a single decorator expression as a string using the IR printer.
-    fn render_single_decorator_expression(&self, dec_idx: NodeIndex) -> Option<String> {
-        use crate::transforms::ir_printer::IRPrinter;
-        let dec_node = self.arena.get(dec_idx)?;
-        let dec = self.arena.get_decorator(dec_node)?;
-        let ir_expr = self.convert_expression_static(dec.expression);
-        let mut printer = IRPrinter::with_arena(self.arena);
-        if let Some(source_text) = self.source_text {
-            printer.set_source_text(source_text);
-        }
-        if let Some(ref transforms) = self.transforms {
-            printer.set_transforms(transforms.clone());
-        }
-        Some(printer.emit(&ir_expr).to_string())
-    }
-
-    /// Render decorator expressions as strings using the IR printer.
-    fn render_decorator_expressions(&self, decorators: &[NodeIndex]) -> Vec<String> {
-        use crate::transforms::ir_printer::IRPrinter;
-        let mut result = Vec::new();
-        for &dec_idx in decorators {
-            if let Some(dec_node) = self.arena.get(dec_idx)
-                && let Some(dec) = self.arena.get_decorator(dec_node)
-            {
-                let ir_expr = self.convert_expression_static(dec.expression);
-                let mut printer = IRPrinter::with_arena(self.arena);
-                if let Some(source_text) = self.source_text {
-                    printer.set_source_text(source_text);
-                }
-                if let Some(ref transforms) = self.transforms {
-                    printer.set_transforms(transforms.clone());
-                }
-                let rendered = printer.emit(&ir_expr).to_string();
-                result.push(rendered);
-            }
-        }
-        result
-    }
-
-    fn collect_tc39_es5_member_decorators(
-        &self,
-        class_data: &tsz_parser::parser::node::ClassData,
-    ) -> Vec<Tc39Es5MemberDecorator> {
-        let mut result = Vec::new();
-        for &member_idx in &class_data.members.nodes {
-            let Some(member_node) = self.arena.get(member_idx) else {
-                continue;
-            };
-
-            let (modifiers, name_idx, kind) = match member_node.kind {
-                k if k == syntax_kind_ext::METHOD_DECLARATION => {
-                    let Some(method) = self.arena.get_method_decl(member_node) else {
-                        continue;
-                    };
-                    if !method.body.is_some() {
-                        continue;
-                    }
-                    (&method.modifiers, method.name, "method")
-                }
-                k if k == syntax_kind_ext::GET_ACCESSOR => {
-                    let Some(accessor) = self.arena.get_accessor(member_node) else {
-                        continue;
-                    };
-                    (&accessor.modifiers, accessor.name, "getter")
-                }
-                k if k == syntax_kind_ext::SET_ACCESSOR => {
-                    let Some(accessor) = self.arena.get_accessor(member_node) else {
-                        continue;
-                    };
-                    (&accessor.modifiers, accessor.name, "setter")
-                }
-                _ => continue,
-            };
-
-            let decorators = self.collect_decorators_from_modifiers(modifiers);
-            if decorators.is_empty() {
-                continue;
-            }
-            let Some(name) = get_identifier_text(self.arena, name_idx) else {
-                continue;
-            };
-            if name.is_empty() {
-                continue;
-            }
-
-            let prefix = if self.arena.is_static(modifiers) {
-                "_static_"
-            } else {
-                "_"
-            };
-            result.push(Tc39Es5MemberDecorator {
-                decorators_var: format!("{prefix}{name}_decorators"),
-                decorator_exprs: self.render_decorator_expressions(&decorators),
-                kind,
-                name,
-                is_static: self.arena.is_static(modifiers),
-            });
-        }
-        result
-    }
-
-    pub fn wrap_tc39_es5_output(
-        &self,
-        class_idx: NodeIndex,
-        override_name: Option<&str>,
-        inner_output: &str,
-    ) -> Option<String> {
-        let class_node = self.arena.get(class_idx)?;
-        let class_data = self.arena.get_class(class_node)?;
-        let class_name = override_name
-            .map(ToOwned::to_owned)
-            .or_else(|| get_identifier_text(self.arena, class_data.name))?;
-        let member_decorators = self.collect_tc39_es5_member_decorators(class_data);
-        if member_decorators.is_empty() {
-            return None;
-        }
-
-        let alias = "_a";
-        let base_indent = "    ".repeat(self.indent_base as usize);
-        let body_indent = "    ".repeat((self.indent_base + 1) as usize);
-        let inner_indent = "    ".repeat((self.indent_base + 2) as usize);
-        let decorator_indent = "    ".repeat((self.indent_base + 3) as usize);
-
-        let prefix = format!("var {class_name} = ");
-        let mut class_expr = inner_output.trim_end().strip_prefix(&prefix)?.to_string();
-        if let Some(stripped) = class_expr.strip_suffix(';') {
-            class_expr = stripped.to_string();
-        }
-        let mut class_expr_lines = class_expr.lines();
-        let first_class_line = class_expr_lines.next().unwrap_or_default();
-
-        let has_instance = member_decorators.iter().any(|member| !member.is_static);
-        let has_static = member_decorators.iter().any(|member| member.is_static);
-
-        let mut out = String::new();
-        out.push_str(&format!("{base_indent}var {class_name} = function () {{\n"));
-        out.push_str(&format!("{body_indent}var {alias};\n"));
-        if has_instance {
-            out.push_str(&format!(
-                "{body_indent}var _instanceExtraInitializers = [];\n"
-            ));
-        }
-        if has_static {
-            out.push_str(&format!(
-                "{body_indent}var _staticExtraInitializers = [];\n"
-            ));
-        }
-        for member in &member_decorators {
-            out.push_str(&format!("{body_indent}var {};\n", member.decorators_var));
-        }
-
-        out.push_str(&format!(
-            "{body_indent}return {alias} = {first_class_line}\n"
-        ));
-        let remaining_class_lines: Vec<&str> = class_expr_lines.collect();
-        for (idx, line) in remaining_class_lines.iter().enumerate() {
-            out.push_str(&inner_indent);
-            out.push_str(line);
-            if idx + 1 == remaining_class_lines.len() {
-                out.push_str(",\n");
-            } else {
-                out.push('\n');
-            }
-        }
-        out.push_str(&format!("{inner_indent}(function () {{\n"));
-        out.push_str(&format!(
-            "{decorator_indent}var _metadata = typeof Symbol === \"function\" && Symbol.metadata ? Object.create(null) : void 0;\n"
-        ));
-        for member in &member_decorators {
-            out.push_str(&format!(
-                "{decorator_indent}{} = [{}];\n",
-                member.decorators_var,
-                member.decorator_exprs.join(", ")
-            ));
-            let extra_var = if member.is_static {
-                "_staticExtraInitializers"
-            } else {
-                "_instanceExtraInitializers"
-            };
-            out.push_str(&format!(
-                "{decorator_indent}__esDecorate({alias}, null, {}, {{ kind: \"{}\", name: \"{}\", static: {}, private: false, access: {{ {} }}, metadata: _metadata }}, null, {extra_var});\n",
-                member.decorators_var,
-                member.kind,
-                member.name,
-                member.is_static,
-                self.tc39_es5_member_access(member),
-            ));
-        }
-        out.push_str(&format!(
-            "{decorator_indent}if (_metadata) Object.defineProperty({alias}, Symbol.metadata, {{ enumerable: true, configurable: true, writable: true, value: _metadata }});\n"
-        ));
-        if has_static {
-            out.push_str(&format!(
-                "{decorator_indent}__runInitializers({alias}, _staticExtraInitializers);\n"
-            ));
-        }
-        out.push_str(&format!("{inner_indent}}})(),\n"));
-        out.push_str(&format!("{inner_indent}{alias};\n"));
-        out.push_str(&format!("{base_indent}}}();"));
-        Some(out)
-    }
-
-    fn tc39_es5_member_access(&self, member: &Tc39Es5MemberDecorator) -> String {
-        let name = &member.name;
-        match member.kind {
-            "setter" => format!(
-                "has: function (obj) {{ return \"{name}\" in obj; }}, set: function (obj, value) {{ obj.{name} = value; }}"
-            ),
-            _ => format!(
-                "has: function (obj) {{ return \"{name}\" in obj; }}, get: function (obj) {{ return obj.{name}; }}"
-            ),
-        }
-    }
-
-    fn accessor_metadata_strings(
-        &self,
-        members: &[NodeIndex],
-        name_idx: NodeIndex,
-        is_static: bool,
-    ) -> Vec<String> {
-        let Some(target_name) = get_identifier_text(self.arena, name_idx) else {
-            return vec![
-                format!(
-                    "{}(\"design:type\", Object)",
-                    self.helper_name("__metadata")
-                ),
-                format!(
-                    "{}(\"design:paramtypes\", [])",
-                    self.helper_name("__metadata")
-                ),
-            ];
-        };
-        let mut setter_parameters: Option<NodeList> = None;
-        let mut getter_type = NodeIndex::NONE;
-
-        for &member_idx in members {
-            let Some(member_node) = self.arena.get(member_idx) else {
-                continue;
-            };
-            if member_node.kind != syntax_kind_ext::GET_ACCESSOR
-                && member_node.kind != syntax_kind_ext::SET_ACCESSOR
-            {
-                continue;
-            }
-            let Some(accessor) = self.arena.get_accessor(member_node) else {
-                continue;
-            };
-            if self.arena.is_static(&accessor.modifiers) != is_static {
-                continue;
-            }
-            if get_identifier_text(self.arena, accessor.name).as_deref() != Some(&target_name) {
-                continue;
-            }
-            if member_node.kind == syntax_kind_ext::SET_ACCESSOR {
-                setter_parameters = Some(accessor.parameters.clone());
-            } else if accessor.type_annotation.is_some() {
-                getter_type = accessor.type_annotation;
-            }
-        }
-
-        let design_type = if let Some(params) = setter_parameters.as_ref() {
-            params
-                .nodes
-                .first()
-                .and_then(|&param_idx| self.arena.get(param_idx))
-                .and_then(|param_node| self.arena.get_parameter(param_node))
-                .and_then(|param| {
-                    param
-                        .type_annotation
-                        .is_some()
-                        .then_some(param.type_annotation)
-                })
-                .map(|type_idx| serialize_type_for_metadata(self.arena, type_idx))
-                .unwrap_or_else(|| "Object".to_string())
-        } else if getter_type.is_some() {
-            serialize_type_for_metadata(self.arena, getter_type)
-        } else {
-            "Object".to_string()
-        };
-
-        let param_types = setter_parameters
-            .as_ref()
-            .map(|params| serialize_param_types(self.arena, params))
-            .unwrap_or_default();
-
-        vec![
-            format!(
-                "{}(\"design:type\", {design_type})",
-                self.helper_name("__metadata")
-            ),
-            format!(
-                "{}(\"design:paramtypes\", [{param_types}])",
-                self.helper_name("__metadata")
-            ),
-        ]
-    }
-
-    /// Emit `__decorate` calls for decorated members inside the IIFE body.
-    fn emit_member_decorator_ir(&self, body: &mut Vec<IRNode>, class_idx: NodeIndex) {
-        let Some(class_node) = self.arena.get(class_idx) else {
-            return;
-        };
-        let Some(class_data) = self.arena.get_class(class_node) else {
-            return;
-        };
-
-        // Track accessor names that have already been emitted so that
-        // getter/setter pairs produce only one __decorate call (the first one).
-        let mut emitted_accessor_names = std::collections::HashSet::<String>::new();
-
-        for &member_idx in &class_data.members.nodes {
-            let Some(member_node) = self.arena.get(member_idx) else {
-                continue;
-            };
-
-            enum MemberMeta {
-                Property {
-                    type_annotation: NodeIndex,
-                },
-                Method {
-                    parameters: NodeList,
-                    return_type: NodeIndex,
-                    async_returns_promise: bool,
-                },
-                Accessor {
-                    name: NodeIndex,
-                    is_static: bool,
-                },
-            }
-
-            let (modifiers, name_idx, is_property, is_accessor, meta) = match member_node.kind {
-                k if k == syntax_kind_ext::METHOD_DECLARATION => {
-                    let Some(method) = self.arena.get_method_decl(member_node) else {
-                        continue;
-                    };
-                    // Skip overload signatures (no body) — decorators on overloads
-                    // are not emitted as __decorate targets
-                    if !method.body.is_some() {
-                        continue;
-                    }
-                    let has_async_modifier = self
-                        .arena
-                        .has_modifier(&method.modifiers, SyntaxKind::AsyncKeyword);
-                    let has_generator_asterisk = method.asterisk_token
-                        || crate::transforms::emit_utils::source_header_has_async_generator_asterisk(
-                            self.source_text,
-                            member_node.pos,
-                            self.arena
-                                .get(method.body)
-                                .map_or(member_node.end, |body| body.pos),
-                        );
-                    let meta = MemberMeta::Method {
-                        parameters: method.parameters.clone(),
-                        return_type: method.type_annotation,
-                        async_returns_promise: has_async_modifier && !has_generator_asterisk,
-                    };
-                    (&method.modifiers, method.name, false, false, meta)
-                }
-                k if k == syntax_kind_ext::PROPERTY_DECLARATION => {
-                    let Some(prop) = self.arena.get_property_decl(member_node) else {
-                        continue;
-                    };
-                    let is_auto_accessor = self
-                        .arena
-                        .has_modifier(&prop.modifiers, SyntaxKind::AccessorKeyword);
-                    let meta = MemberMeta::Property {
-                        type_annotation: prop.type_annotation,
-                    };
-                    (&prop.modifiers, prop.name, !is_auto_accessor, false, meta)
-                }
-                k if k == syntax_kind_ext::GET_ACCESSOR || k == syntax_kind_ext::SET_ACCESSOR => {
-                    let Some(accessor) = self.arena.get_accessor(member_node) else {
-                        continue;
-                    };
-                    (
-                        &accessor.modifiers,
-                        accessor.name,
-                        false,
-                        true,
-                        MemberMeta::Accessor {
-                            name: accessor.name,
-                            is_static: self.arena.is_static(&accessor.modifiers),
-                        },
-                    )
-                }
-                _ => continue,
-            };
-
-            let decorators = self.collect_decorators_from_modifiers(modifiers);
-
-            // Collect parameter decorators for methods/constructors.
-            // Each entry is (runtime_param_index, decorator_nodes).
-            let param_decorators: Vec<(usize, Vec<NodeIndex>)> = match &meta {
-                MemberMeta::Method { parameters, .. } => {
-                    self.collect_param_decorators_es5(parameters)
-                }
-                _ => Vec::new(),
-            };
-
-            if decorators.is_empty() && param_decorators.is_empty() {
-                continue;
-            }
-
-            let is_static = self.arena.is_static(modifiers);
-
-            let member_name = get_identifier_text(self.arena, name_idx);
-            let Some(member_name) = member_name else {
-                continue;
-            };
-            if member_name.is_empty() {
-                continue;
-            }
-
-            // For getter/setter pairs, tsc emits only one __decorate call
-            // for the first accessor that has decorators. Skip the second.
-            if is_accessor && !emitted_accessor_names.insert(member_name.clone()) {
-                continue;
-            }
-
-            let mut dec_strs = self.render_decorator_expressions(&decorators);
-            // Add __param entries for parameter decorators
-            for (param_idx, param_decs) in &param_decorators {
-                for dec_idx in param_decs {
-                    let dec_str = self.render_single_decorator_expression(*dec_idx);
-                    if let Some(dec_str) = dec_str {
-                        dec_strs.push(format!(
-                            "{}({param_idx}, {dec_str})",
-                            self.helper_name("__param")
-                        ));
-                    }
-                }
-            }
-            let target_str = if is_static {
-                self.class_name.clone()
-            } else {
-                format!("{}.prototype", self.class_name)
-            };
-            let desc_str = if is_property { "void 0" } else { "null" };
-
-            // Collect metadata strings if emit_decorator_metadata is enabled
-            let metadata_strs: Vec<String> = if self.emit_decorator_metadata {
-                match &meta {
-                    MemberMeta::Property { type_annotation } => {
-                        let serialized = serialize_type_for_metadata(self.arena, *type_annotation);
-                        vec![format!(
-                            "{}(\"design:type\", {serialized})",
-                            self.helper_name("__metadata")
-                        )]
-                    }
-                    MemberMeta::Method {
-                        parameters,
-                        return_type,
-                        async_returns_promise,
-                    } => {
-                        let param_types = serialize_param_types(self.arena, parameters);
-                        let ret_type = if return_type.is_some() {
-                            serialize_type_for_metadata(self.arena, *return_type)
-                        } else if *async_returns_promise {
-                            "Promise".to_string()
-                        } else {
-                            "void 0".to_string()
-                        };
-                        vec![
-                            format!(
-                                "{}(\"design:type\", Function)",
-                                self.helper_name("__metadata")
-                            ),
-                            format!(
-                                "{}(\"design:paramtypes\", [{param_types}])",
-                                self.helper_name("__metadata")
-                            ),
-                            format!(
-                                "{}(\"design:returntype\", {ret_type})",
-                                self.helper_name("__metadata")
-                            ),
-                        ]
-                    }
-                    MemberMeta::Accessor { name, is_static } => {
-                        self.accessor_metadata_strings(&class_data.members.nodes, *name, *is_static)
-                    }
-                }
-            } else {
-                Vec::new()
-            };
-
-            // Format matching tsc:
-            // __decorate([\n        dec1,\n        dec2\n    ], target, "name", desc)
-            // Note: first line indent is handled by the body emitter's write_indent().
-            // Continuation lines after \n need absolute indentation from column 0.
-            // The indent_base accounts for nesting (e.g., namespace IIFE body).
-            let inner_indent = "    ".repeat((self.indent_base + 2) as usize);
-            let outer_indent = "    ".repeat((self.indent_base + 1) as usize);
-            let total_entries = dec_strs.len() + metadata_strs.len();
-            let mut raw = String::new();
-            raw.push_str(&self.helper_name("__decorate"));
-            raw.push_str("([");
-            for (i, dec_str) in dec_strs.iter().enumerate() {
-                raw.push('\n');
-                raw.push_str(&inner_indent);
-                raw.push_str(dec_str);
-                if i + 1 < total_entries {
-                    raw.push(',');
-                }
-            }
-            for (i, meta_str) in metadata_strs.iter().enumerate() {
-                raw.push('\n');
-                raw.push_str(&inner_indent);
-                raw.push_str(meta_str);
-                if dec_strs.len() + i + 1 < total_entries {
-                    raw.push(',');
-                }
-            }
-            raw.push('\n');
-            raw.push_str(&outer_indent);
-            raw.push_str("], ");
-            raw.push_str(&target_str);
-            raw.push_str(", \"");
-            raw.push_str(&member_name);
-            raw.push_str("\", ");
-            raw.push_str(desc_str);
-            raw.push(')');
-
-            body.push(IRNode::ExpressionStatement(Box::new(IRNode::Raw(
-                raw.into(),
-            ))));
-        }
-    }
-
-    /// Emit `ClassName = __decorate([dec1, ...], ClassName)` for class-level decorators.
-    /// When `emit_decorator_metadata` is enabled and the class has a constructor,
-    /// also includes `__metadata("design:paramtypes", [...])` in the decorator array.
-    fn emit_class_decorator_ir(&self, body: &mut Vec<IRNode>, class_idx: NodeIndex) {
-        let dec_strs = self.render_decorator_expressions(&self.class_decorators);
-        if dec_strs.is_empty() {
-            return;
-        }
-
-        // Collect constructor parameter decorators (__param entries).
-        // tsc includes these in the class-level __decorate call between
-        // class decorators and __metadata entries.
-        let mut param_strs: Vec<String> = Vec::new();
-        let mut metadata_strs: Vec<String> = Vec::new();
-        if let Some(class_node) = self.arena.get(class_idx)
-            && let Some(class_data) = self.arena.get_class(class_node)
-        {
-            for &member_idx in &class_data.members.nodes {
-                if let Some(member_node) = self.arena.get(member_idx)
-                    && member_node.kind == syntax_kind_ext::CONSTRUCTOR
-                    && let Some(ctor) = self.arena.get_constructor(member_node)
-                {
-                    // Collect __param entries for constructor parameter decorators
-                    let all_param_decs = self.collect_param_decorators_es5(&ctor.parameters);
-                    for (param_idx, decs) in &all_param_decs {
-                        for dec_idx in decs {
-                            if let Some(dec_str) = self.render_single_decorator_expression(*dec_idx)
-                            {
-                                param_strs.push(format!(
-                                    "{}({param_idx}, {dec_str})",
-                                    self.helper_name("__param")
-                                ));
-                            }
-                        }
-                    }
-
-                    // Build constructor paramtypes metadata if emit_decorator_metadata is enabled
-                    if self.emit_decorator_metadata {
-                        let param_types = serialize_param_types(self.arena, &ctor.parameters);
-                        metadata_strs.push(format!(
-                            "{}(\"design:paramtypes\", [{param_types}])",
-                            self.helper_name("__metadata")
-                        ));
-                    }
-                    break;
-                }
-            }
-        }
-
-        // Format matching tsc:
-        // ClassName = __decorate([\n        dec1,\n        __param(0, dec),\n        __metadata(...)\n    ], ClassName)
-        let inner_indent = "    ".repeat((self.indent_base + 2) as usize);
-        let outer_indent = "    ".repeat((self.indent_base + 1) as usize);
-        let total_entries = dec_strs.len() + param_strs.len() + metadata_strs.len();
-        let mut raw = String::new();
-        raw.push_str(&self.class_name);
-        raw.push_str(" = ");
-        if let Some(alias) = self.class_self_reference_alias.as_ref() {
-            raw.push_str(alias);
-            raw.push_str(" = ");
-        }
-        raw.push_str(&self.helper_name("__decorate"));
-        raw.push_str("([");
-        let mut written = 0;
-        for dec_str in &dec_strs {
-            raw.push('\n');
-            raw.push_str(&inner_indent);
-            raw.push_str(dec_str);
-            written += 1;
-            if written < total_entries {
-                raw.push(',');
-            }
-        }
-        for param_str in &param_strs {
-            raw.push('\n');
-            raw.push_str(&inner_indent);
-            raw.push_str(param_str);
-            written += 1;
-            if written < total_entries {
-                raw.push(',');
-            }
-        }
-        for meta_str in &metadata_strs {
-            raw.push('\n');
-            raw.push_str(&inner_indent);
-            raw.push_str(meta_str);
-            written += 1;
-            if written < total_entries {
-                raw.push(',');
-            }
-        }
-        raw.push('\n');
-        raw.push_str(&outer_indent);
-        raw.push_str("], ");
-        raw.push_str(&self.class_name);
-        raw.push(')');
-
-        body.push(IRNode::ExpressionStatement(Box::new(IRNode::Raw(
-            raw.into(),
-        ))));
-    }
-
-    /// Emit `ClassName = __decorate([__param(0, dec), ...], ClassName)` for constructor
-    /// parameter decorators when there are no class-level decorators. tsc emits this
-    /// at the class level when a constructor parameter has a decorator.
-    fn emit_ctor_param_decorator_ir(&self, body: &mut Vec<IRNode>, class_idx: NodeIndex) {
-        let Some(class_node) = self.arena.get(class_idx) else {
-            return;
-        };
-        let Some(class_data) = self.arena.get_class(class_node) else {
-            return;
-        };
-
-        // Find the constructor and collect its parameter decorators
-        let mut all_param_decs: Vec<(usize, Vec<NodeIndex>)> = Vec::new();
-        for &member_idx in &class_data.members.nodes {
-            if let Some(member_node) = self.arena.get(member_idx)
-                && member_node.kind == syntax_kind_ext::CONSTRUCTOR
-                && let Some(ctor) = self.arena.get_constructor(member_node)
-            {
-                all_param_decs = self.collect_param_decorators_es5(&ctor.parameters);
-                break;
-            }
-        }
-
-        if all_param_decs.is_empty() {
-            return;
-        }
-
-        // Build __param(index, dec) strings
-        let mut param_strs: Vec<String> = Vec::new();
-        for (param_idx, decs) in &all_param_decs {
-            for dec_idx in decs {
-                if let Some(dec_str) = self.render_single_decorator_expression(*dec_idx) {
-                    param_strs.push(format!(
-                        "{}({param_idx}, {dec_str})",
-                        self.helper_name("__param")
-                    ));
-                }
-            }
-        }
-
-        if param_strs.is_empty() {
-            return;
-        }
-
-        // Build constructor paramtypes metadata if emit_decorator_metadata is enabled
-        let metadata_strs: Vec<String> = if self.emit_decorator_metadata {
-            let mut meta = Vec::new();
-            for &member_idx in &class_data.members.nodes {
-                if let Some(member_node) = self.arena.get(member_idx)
-                    && member_node.kind == syntax_kind_ext::CONSTRUCTOR
-                    && let Some(ctor) = self.arena.get_constructor(member_node)
-                {
-                    let param_types = serialize_param_types(self.arena, &ctor.parameters);
-                    meta.push(format!(
-                        "{}(\"design:paramtypes\", [{param_types}])",
-                        self.helper_name("__metadata")
-                    ));
-                    break;
-                }
-            }
-            meta
-        } else {
-            Vec::new()
-        };
-
-        let inner_indent = "    ".repeat((self.indent_base + 2) as usize);
-        let outer_indent = "    ".repeat((self.indent_base + 1) as usize);
-        let total_entries = param_strs.len() + metadata_strs.len();
-        let mut raw = String::new();
-        raw.push_str(&self.class_name);
-        raw.push_str(" = ");
-        raw.push_str(&self.helper_name("__decorate"));
-        raw.push_str("([");
-        for (i, param_str) in param_strs.iter().enumerate() {
-            raw.push('\n');
-            raw.push_str(&inner_indent);
-            raw.push_str(param_str);
-            if i + 1 < total_entries {
-                raw.push(',');
-            }
-        }
-        for (i, meta_str) in metadata_strs.iter().enumerate() {
-            raw.push('\n');
-            raw.push_str(&inner_indent);
-            raw.push_str(meta_str);
-            if param_strs.len() + i + 1 < total_entries {
-                raw.push(',');
-            }
-        }
-        raw.push('\n');
-        raw.push_str(&outer_indent);
-        raw.push_str("], ");
-        raw.push_str(&self.class_name);
-        raw.push(')');
-
-        body.push(IRNode::ExpressionStatement(Box::new(IRNode::Raw(
-            raw.into(),
-        ))));
-    }
-
     /// Convert a block body to IR statements
     fn convert_block_body(&self, block_idx: NodeIndex) -> Vec<IRNode> {
         self.convert_block_body_with_alias(block_idx, None)
@@ -2564,881 +1775,6 @@ impl<'a> ES5ClassTransformer<'a> {
             deferred_static_blocks,
             deferred_block_class_alias,
         })
-    }
-
-    fn auto_accessor_has_computed_name(&self, member_idx: NodeIndex) -> bool {
-        let Some(member_node) = self.arena.get(member_idx) else {
-            return false;
-        };
-        let Some(prop) = self.arena.get_property_decl(member_node) else {
-            return false;
-        };
-        self.arena
-            .get(prop.name)
-            .is_some_and(|name| name.kind == syntax_kind_ext::COMPUTED_PROPERTY_NAME)
-    }
-
-    fn auto_accessor_instance_storage_inits_for_computed_key(
-        &self,
-        member_idx: NodeIndex,
-    ) -> Vec<String> {
-        if self.first_computed_instance_auto_accessor().is_none()
-            || self
-                .first_computed_instance_auto_accessor()
-                .is_some_and(|accessor| accessor.member_idx != member_idx)
-        {
-            return Vec::new();
-        }
-
-        self.auto_accessors
-            .iter()
-            .filter(|accessor| !accessor.is_static)
-            .map(|accessor| format!("{} = new WeakMap()", accessor.weakmap_name))
-            .collect()
-    }
-
-    fn emit_auto_accessor_storage_decls_and_static_inits(&self, body: &mut Vec<IRNode>) {
-        let mut names = Vec::new();
-        if let Some(alias) = self.current_static_class_alias.as_ref() {
-            names.push(alias.clone());
-        }
-        names.extend(
-            self.auto_accessors
-                .iter()
-                .map(|accessor| accessor.weakmap_name.clone()),
-        );
-        if !names.is_empty() {
-            body.push(IRNode::VarDeclList(
-                names
-                    .into_iter()
-                    .map(|name| IRNode::VarDecl {
-                        name: name.into(),
-                        initializer: None,
-                    })
-                    .collect(),
-            ));
-        }
-
-        if let Some(alias) = self.current_static_class_alias.as_ref() {
-            body.push(IRNode::expr_stmt(IRNode::assign(
-                IRNode::id(alias.clone()),
-                IRNode::id(self.class_name.clone()),
-            )));
-        }
-
-        if self.first_computed_instance_auto_accessor().is_none() {
-            for accessor in &self.auto_accessors {
-                if accessor.is_static {
-                    continue;
-                }
-                body.push(IRNode::expr_stmt(IRNode::assign(
-                    IRNode::id(accessor.weakmap_name.clone()),
-                    IRNode::NewExpr {
-                        callee: Box::new(IRNode::id("WeakMap")),
-                        arguments: Vec::new(),
-                        explicit_arguments: true,
-                    },
-                )));
-            }
-        }
-
-        for accessor in &self.auto_accessors {
-            if !accessor.is_static {
-                continue;
-            }
-            let value = accessor
-                .initializer
-                .map(|initializer| self.convert_expression_static(initializer))
-                .unwrap_or(IRNode::Undefined);
-            body.push(IRNode::expr_stmt(IRNode::assign(
-                IRNode::id(accessor.weakmap_name.clone()),
-                IRNode::object(vec![IRProperty::init("value", value)]),
-            )));
-        }
-    }
-
-    fn build_auto_accessor_getter_function(&self, weakmap_name: &str) -> IRNode {
-        IRNode::FunctionExpr {
-            name: None,
-            parameters: vec![],
-            body: vec![IRNode::ret(Some(IRNode::PrivateFieldGet {
-                receiver: Box::new(IRNode::this()),
-                weakmap_name: weakmap_name.to_string().into(),
-            }))],
-            is_expression_body: true,
-            body_source_range: None,
-        }
-    }
-
-    fn build_static_auto_accessor_getter_function(&self, weakmap_name: &str) -> IRNode {
-        let class_alias = self
-            .current_static_class_alias
-            .as_ref()
-            .cloned()
-            .unwrap_or_else(|| self.class_name.clone());
-        IRNode::FunctionExpr {
-            name: None,
-            parameters: vec![],
-            body: vec![IRNode::ret(Some(IRNode::PrivateStaticFieldGet {
-                receiver: Box::new(IRNode::id(class_alias.clone())),
-                state: Box::new(IRNode::id(class_alias)),
-                storage_name: weakmap_name.to_string().into(),
-            }))],
-            is_expression_body: true,
-            body_source_range: None,
-        }
-    }
-
-    fn build_auto_accessor_setter_function(&self, weakmap_name: &str) -> IRNode {
-        IRNode::FunctionExpr {
-            name: None,
-            parameters: vec![IRParam::new("value")],
-            body: vec![IRNode::expr_stmt(IRNode::PrivateFieldSet {
-                receiver: Box::new(IRNode::this()),
-                weakmap_name: weakmap_name.to_string().into(),
-                value: Box::new(IRNode::id("value")),
-            })],
-            is_expression_body: true,
-            body_source_range: None,
-        }
-    }
-
-    fn build_static_auto_accessor_setter_function(&self, weakmap_name: &str) -> IRNode {
-        let class_alias = self
-            .current_static_class_alias
-            .as_ref()
-            .cloned()
-            .unwrap_or_else(|| self.class_name.clone());
-        IRNode::FunctionExpr {
-            name: None,
-            parameters: vec![IRParam::new("value")],
-            body: vec![IRNode::expr_stmt(IRNode::PrivateStaticFieldSet {
-                receiver: Box::new(IRNode::id(class_alias.clone())),
-                state: Box::new(IRNode::id(class_alias)),
-                storage_name: weakmap_name.to_string().into(),
-                value: Box::new(IRNode::id("value")),
-            })],
-            is_expression_body: true,
-            body_source_range: None,
-        }
-    }
-
-    fn auto_accessor_getter_property_name(
-        &self,
-        name_idx: NodeIndex,
-        storage_inits: &[String],
-    ) -> IRMethodName {
-        if storage_inits.is_empty() {
-            return self.auto_accessor_setter_property_name(name_idx);
-        }
-        let Some(name_node) = self.arena.get(name_idx) else {
-            return self.get_method_name_ir(name_idx);
-        };
-        if name_node.kind != syntax_kind_ext::COMPUTED_PROPERTY_NAME {
-            return self.get_method_name_ir(name_idx);
-        }
-        let Some(computed) = self.arena.get_computed_property(name_node) else {
-            return self.get_method_name_ir(name_idx);
-        };
-
-        let expr = self.convert_computed_property_expression(computed.expression, true);
-        let expr_text = self.render_ir_expression(&expr);
-        let mut parts = storage_inits.to_vec();
-        if let Some(temp) = self.computed_prop_temp_map.get(&computed.expression) {
-            parts.push(format!("{temp} = {expr_text}"));
-        } else {
-            parts.push(expr_text);
-        }
-        IRMethodName::Computed(Box::new(IRNode::Raw(
-            format!("({})", parts.join(", ")).into(),
-        )))
-    }
-
-    fn auto_accessor_setter_property_name(&self, name_idx: NodeIndex) -> IRMethodName {
-        let Some(name_node) = self.arena.get(name_idx) else {
-            return self.get_method_name_ir(name_idx);
-        };
-        if name_node.kind != syntax_kind_ext::COMPUTED_PROPERTY_NAME {
-            return self.get_method_name_ir(name_idx);
-        }
-        let Some(computed) = self.arena.get_computed_property(name_node) else {
-            return self.get_method_name_ir(name_idx);
-        };
-        if let Some(temp) = self.computed_prop_temp_map.get(&computed.expression) {
-            return IRMethodName::Computed(Box::new(IRNode::id(temp.clone())));
-        }
-        self.get_method_name_ir(name_idx)
-    }
-
-    pub(super) fn get_field_define_property_name_ir(&self, name_idx: NodeIndex) -> IRMethodName {
-        let Some(name_node) = self.arena.get(name_idx) else {
-            return self.get_method_name_ir(name_idx);
-        };
-        if name_node.kind != syntax_kind_ext::COMPUTED_PROPERTY_NAME {
-            return self.get_method_name_ir(name_idx);
-        }
-        let Some(computed) = self.arena.get_computed_property(name_node) else {
-            return self.get_method_name_ir(name_idx);
-        };
-        if let Some(temp) = self.computed_prop_temp_map.get(&computed.expression) {
-            return IRMethodName::Computed(Box::new(IRNode::id(temp.clone())));
-        }
-        self.get_method_name_ir(name_idx)
-    }
-
-    fn render_ir_expression(&self, expr: &IRNode) -> String {
-        let mut printer = IRPrinter::with_arena(self.arena);
-        printer.set_target_es5(true);
-        if let Some(source_text) = self.source_text {
-            printer.set_source_text(source_text);
-        }
-        if let Some(transforms) = self.transforms.as_ref() {
-            printer.set_transforms(transforms.clone());
-        }
-        printer.emit(expr).to_string()
-    }
-
-    /// Emit a property initializer as an assignment or defineProperty.
-    fn emit_property_initializer_ir(&self, prop_idx: NodeIndex, use_this: bool) -> Option<IRNode> {
-        let prop_node = self.arena.get(prop_idx)?;
-        let prop_data = self.arena.get_property_decl(prop_node)?;
-
-        let has_initializer_equals = self.property_initializer_has_equals(prop_node, prop_data);
-        if !self.use_define_for_class_fields && !has_initializer_equals {
-            return None;
-        }
-
-        let receiver = if use_this {
-            IRNode::id("_this")
-        } else {
-            IRNode::this()
-        };
-
-        let prop_name = self.get_property_name_ir(prop_data.name)?;
-
-        let value = if has_initializer_equals {
-            self.convert_async_arrow_property_initializer(prop_data.initializer)
-                .unwrap_or_else(|| {
-                    if use_this {
-                        self.convert_expression_this_captured(prop_data.initializer)
-                    } else {
-                        self.convert_expression(prop_data.initializer)
-                    }
-                })
-        } else {
-            IRNode::void_0()
-        };
-
-        if self.use_define_for_class_fields {
-            Some(IRNode::DefineProperty {
-                target: Box::new(receiver),
-                property_name: self.get_field_define_property_name_ir(prop_data.name),
-                descriptor: IRPropertyDescriptor {
-                    get: None,
-                    set: None,
-                    value: Some(Box::new(value)),
-                    get_leading_comment: None,
-                    set_leading_comment: None,
-                    enumerable: true,
-                    configurable: true,
-                    writable: true,
-                    trailing_comment: None,
-                },
-                leading_comment: None,
-            })
-        } else {
-            Some(IRNode::expr_stmt(IRNode::assign(
-                self.build_property_access(receiver, prop_name),
-                value,
-            )))
-        }
-    }
-
-    /// Build property access node based on property name type
-    fn build_property_access(&self, receiver: IRNode, name: PropertyNameIR) -> IRNode {
-        match name {
-            PropertyNameIR::Identifier(n) => IRNode::prop(receiver, n),
-            PropertyNameIR::StringLiteral(s) => IRNode::elem(receiver, IRNode::string(s)),
-            PropertyNameIR::NumericLiteral(n) => IRNode::elem(receiver, IRNode::number(n)),
-            PropertyNameIR::Computed(expr_idx) => {
-                // If this expression has a hoisted temp variable, use it
-                if let Some(temp) = self.computed_prop_temp_map.get(&expr_idx) {
-                    IRNode::elem(receiver, IRNode::id(temp.clone()))
-                } else {
-                    IRNode::elem(
-                        receiver,
-                        self.convert_computed_property_expression(expr_idx, false),
-                    )
-                }
-            }
-        }
-    }
-
-    fn convert_async_arrow_property_initializer(&self, initializer: NodeIndex) -> Option<IRNode> {
-        let node = self.arena.get(initializer)?;
-        if node.kind != syntax_kind_ext::ARROW_FUNCTION {
-            return None;
-        }
-        let arrow = self.arena.get_function(node)?;
-        if !arrow.is_async {
-            return None;
-        }
-
-        let mut async_transformer = AsyncES5Transformer::new(self.arena);
-        if let Some(source_text) = self.source_text {
-            async_transformer.set_source_text(source_text);
-        }
-        async_transformer.set_module_kind(self.module_kind);
-        self.configure_async_disposable_context(&mut async_transformer);
-        let has_await = async_transformer.body_contains_await(arrow.body);
-        let mut generator_body = async_transformer.transform_generator_body(arrow.body, has_await);
-        self.sync_async_disposable_context(&mut async_transformer);
-        let hoisted_var_groups =
-            AsyncES5Transformer::extract_and_remove_var_decl_groups(&mut generator_body);
-
-        Some(IRNode::FunctionExpr {
-            name: None,
-            parameters: self.extract_parameters(&arrow.parameters),
-            body: vec![IRNode::AwaiterCall {
-                this_arg: Box::new(IRNode::id("_this")),
-                needs_lexical_this_capture: generator_body.contains_captured_this_reference(),
-                generator_body: Box::new(generator_body),
-                hoisted_var_groups,
-                promise_constructor: self.async_method_promise_constructor(arrow.type_annotation),
-                multiline_callback: false,
-                directives: Vec::new(),
-            }],
-            is_expression_body: true,
-            body_source_range: None,
-        })
-    }
-
-    /// Get property name as IR-friendly representation
-    fn get_property_name_ir(&self, name_idx: NodeIndex) -> Option<PropertyNameIR> {
-        let name_node = self.arena.get(name_idx)?;
-
-        if name_node.kind == syntax_kind_ext::COMPUTED_PROPERTY_NAME {
-            if let Some(computed) = self.arena.get_computed_property(name_node) {
-                return Some(PropertyNameIR::Computed(computed.expression));
-            }
-        } else if name_node.kind == SyntaxKind::Identifier as u16 {
-            if let Some(ident) = self.arena.get_identifier(name_node) {
-                return Some(PropertyNameIR::Identifier(ident.escaped_text.clone()));
-            }
-        } else if name_node.kind == SyntaxKind::StringLiteral as u16 {
-            if let Some(lit) = self.arena.get_literal(name_node) {
-                return Some(PropertyNameIR::StringLiteral(lit.text.clone()));
-            }
-        } else if name_node.kind == SyntaxKind::NumericLiteral as u16
-            && let Some(lit) = self.arena.get_literal(name_node)
-        {
-            return Some(PropertyNameIR::NumericLiteral(lit.text.clone()));
-        }
-
-        None
-    }
-
-    /// Extract parameters from a parameter list
-    fn extract_parameters(&self, params: &NodeList) -> Vec<IRParam> {
-        let mut result = Vec::new();
-        let mut temp_counter: u8 = b'a';
-
-        for &param_idx in &params.nodes {
-            let Some(param_node) = self.arena.get(param_idx) else {
-                continue;
-            };
-            let Some(param) = self.arena.get_parameter(param_node) else {
-                continue;
-            };
-
-            // Skip `this` parameter — it's TypeScript-only and erased in JS emit.
-            // The parser may store it as an Identifier with text "this" or as a ThisKeyword token.
-            if let Some(name_node) = self.arena.get(param.name)
-                && name_node.kind == SyntaxKind::ThisKeyword as u16
-            {
-                continue;
-            }
-
-            let mut name = get_identifier_text(self.arena, param.name).unwrap_or_default();
-            if name == "this" {
-                continue;
-            }
-            // For destructured parameters (binding patterns), generate a temp name
-            if name.is_empty() {
-                let name_node = self.arena.get(param.name);
-                let is_binding_pattern = name_node.is_some_and(|n| {
-                    n.kind == syntax_kind_ext::OBJECT_BINDING_PATTERN
-                        || n.kind == syntax_kind_ext::ARRAY_BINDING_PATTERN
-                });
-                if is_binding_pattern {
-                    name = format!("_{}", temp_counter as char);
-                    temp_counter = temp_counter.wrapping_add(1);
-                } else {
-                    continue;
-                }
-            }
-
-            let is_rest = param.dot_dot_dot_token;
-            let mut ir_param = if is_rest {
-                IRParam::rest(name)
-            } else {
-                IRParam::new(name)
-            };
-
-            // Convert default value if present
-            if param.initializer.is_some() {
-                ir_param.default_value = Some(Box::new(self.convert_expression(param.initializer)));
-            }
-            if let Some(name_node) = self.arena.get(param.name)
-                && let Some(comment) = self.extract_leading_comment(name_node)
-            {
-                ir_param.leading_comment = Some(comment.into());
-            }
-
-            result.push(ir_param);
-        }
-
-        result
-    }
-
-    /// Generate destructuring prologue IR nodes for binding-pattern parameters.
-    /// For `({ a, b })` with temp name `_a`, generates: `var a = _a.a, b = _a.b;`
-    fn generate_destructuring_prologue(
-        &self,
-        ast_params: &tsz_parser::parser::NodeList,
-        ir_params: &[IRParam],
-    ) -> Vec<IRNode> {
-        let mut prologue = Vec::new();
-        let mut ir_idx = 0;
-        let mut reserved_temp_names: FxHashSet<String> = ir_params
-            .iter()
-            .map(|param| param.name.to_string())
-            .collect();
-
-        for &param_idx in &ast_params.nodes {
-            let Some(param_node) = self.arena.get(param_idx) else {
-                ir_idx += 1;
-                continue;
-            };
-            let Some(param) = self.arena.get_parameter(param_node) else {
-                ir_idx += 1;
-                continue;
-            };
-
-            let name_node = self.arena.get(param.name);
-
-            // Skip `this` parameter — it was also skipped in extract_parameters,
-            // so don't increment ir_idx.
-            let is_this = name_node.is_some_and(|n| n.kind == SyntaxKind::ThisKeyword as u16)
-                || get_identifier_text(self.arena, param.name).as_deref() == Some("this");
-            if is_this {
-                continue;
-            }
-
-            let is_binding_pattern = name_node.is_some_and(|n| {
-                n.kind == syntax_kind_ext::OBJECT_BINDING_PATTERN
-                    || n.kind == syntax_kind_ext::ARRAY_BINDING_PATTERN
-            });
-
-            if !is_binding_pattern {
-                ir_idx += 1;
-                continue;
-            }
-
-            // Get the temp name from the corresponding IR param
-            let temp_name = if ir_idx < ir_params.len() {
-                ir_params[ir_idx].name.to_string()
-            } else {
-                ir_idx += 1;
-                continue;
-            };
-
-            if let Some(name_n) = name_node
-                && name_n.kind == syntax_kind_ext::OBJECT_BINDING_PATTERN
-                && let Some(pattern) = self.arena.get_binding_pattern(name_n)
-            {
-                let mut declarations = Vec::new();
-                let mut rest_excluded = Vec::new();
-                for &elem_idx in &pattern.elements.nodes {
-                    if let Some(elem_node) = self.arena.get(elem_idx)
-                        && let Some(elem) = self.arena.get_binding_element(elem_node)
-                    {
-                        let elem_name =
-                            get_identifier_text(self.arena, elem.name).unwrap_or_default();
-                        if !elem_name.is_empty() {
-                            if elem.dot_dot_dot_token {
-                                let excluded =
-                                    rest_excluded.iter().cloned().map(IRNode::string).collect();
-                                declarations.push(IRNode::var_decl(
-                                    elem_name,
-                                    Some(IRNode::call(
-                                        IRNode::RuntimeHelper("__rest".into()),
-                                        vec![
-                                            IRNode::id(temp_name.clone()),
-                                            IRNode::ArrayLiteral(excluded),
-                                        ],
-                                    )),
-                                ));
-                                continue;
-                            }
-
-                            let prop_name = if elem.property_name.is_some() {
-                                get_identifier_text(self.arena, elem.property_name)
-                                    .unwrap_or_else(|| elem_name.clone())
-                            } else {
-                                elem_name.clone()
-                            };
-                            rest_excluded.push(prop_name.clone());
-                            declarations.push(IRNode::var_decl(
-                                elem_name,
-                                Some(IRNode::prop(IRNode::id(temp_name.clone()), prop_name)),
-                            ));
-                        }
-                    }
-                }
-                if !declarations.is_empty() {
-                    prologue.push(IRNode::VarDeclList(declarations));
-                }
-            } else if let Some(name_n) = name_node
-                && name_n.kind == syntax_kind_ext::ARRAY_BINDING_PATTERN
-                && let Some(pattern) = self.arena.get_binding_pattern(name_n)
-            {
-                let mut declarations = Vec::new();
-                let source_name = if self.downlevel_iteration && !pattern.elements.nodes.is_empty()
-                {
-                    let read_name = Self::fresh_destructuring_temp(&mut reserved_temp_names);
-                    let mut read_args = vec![IRNode::id(temp_name.clone())];
-                    if let Some(limit) = self.array_binding_read_limit(pattern) {
-                        read_args.push(IRNode::number(limit.to_string()));
-                    }
-                    declarations.push(IRNode::var_decl(
-                        read_name.clone(),
-                        Some(IRNode::call(
-                            IRNode::RuntimeHelper("__read".into()),
-                            read_args,
-                        )),
-                    ));
-                    read_name
-                } else {
-                    temp_name.clone()
-                };
-
-                for (element_index, &elem_idx) in pattern.elements.nodes.iter().enumerate() {
-                    let Some(elem_node) = self.arena.get(elem_idx) else {
-                        continue;
-                    };
-                    let Some(elem) = self.arena.get_binding_element(elem_node) else {
-                        continue;
-                    };
-                    let elem_name = get_identifier_text(self.arena, elem.name).unwrap_or_default();
-                    if elem_name.is_empty() {
-                        continue;
-                    }
-
-                    let initializer = if elem.dot_dot_dot_token {
-                        IRNode::call(
-                            IRNode::prop(IRNode::id(source_name.clone()), "slice"),
-                            vec![IRNode::number(element_index.to_string())],
-                        )
-                    } else {
-                        IRNode::elem(
-                            IRNode::id(source_name.clone()),
-                            IRNode::number(element_index.to_string()),
-                        )
-                    };
-                    declarations.push(IRNode::var_decl(elem_name, Some(initializer)));
-                }
-
-                if !declarations.is_empty() {
-                    prologue.push(IRNode::VarDeclList(declarations));
-                }
-            }
-            ir_idx += 1;
-        }
-        prologue
-    }
-
-    fn fresh_destructuring_temp(reserved: &mut FxHashSet<String>) -> String {
-        let mut idx = 0usize;
-        loop {
-            let candidate = if idx < 26 {
-                format!("_{}", (b'a' + idx as u8) as char)
-            } else {
-                format!("_{idx}")
-            };
-            if reserved.insert(candidate.clone()) {
-                return candidate;
-            }
-            idx += 1;
-        }
-    }
-
-    fn array_binding_read_limit(
-        &self,
-        pattern: &tsz_parser::parser::node::BindingPatternData,
-    ) -> Option<usize> {
-        for &elem_idx in &pattern.elements.nodes {
-            if self
-                .arena
-                .get(elem_idx)
-                .and_then(|node| self.arena.get_binding_element(node))
-                .is_some_and(|elem| elem.dot_dot_dot_token)
-            {
-                return None;
-            }
-        }
-        Some(pattern.elements.nodes.len())
-    }
-
-    /// Check if any parameters are destructured binding patterns.
-    pub(super) fn has_destructured_parameters(
-        &self,
-        params: &tsz_parser::parser::NodeList,
-    ) -> bool {
-        params.nodes.iter().any(|&param_idx| {
-            self.arena
-                .get(param_idx)
-                .and_then(|n| self.arena.get_parameter(n))
-                .and_then(|p| self.arena.get(p.name))
-                .is_some_and(|n| {
-                    n.kind == syntax_kind_ext::OBJECT_BINDING_PATTERN
-                        || n.kind == syntax_kind_ext::ARRAY_BINDING_PATTERN
-                })
-        })
-    }
-
-    /// Get the extends clause base class
-    fn get_extends_class(&self, heritage_clauses: &Option<NodeList>) -> Option<IRNode> {
-        let expr_idx = crate::transforms::emit_utils::get_extends_expression_index(
-            self.arena,
-            heritage_clauses,
-        )?;
-        Some(self.convert_expression(expr_idx))
-    }
-
-    /// Collect all arrow function node indices in a block
-    fn collect_arrow_functions_in_block(&self, block_idx: NodeIndex) -> Vec<NodeIndex> {
-        let mut arrows = Vec::new();
-        if let Some(block_node) = self.arena.get(block_idx)
-            && let Some(block) = self.arena.get_block(block_node)
-        {
-            for &stmt_idx in &block.statements.nodes {
-                self.collect_arrow_functions_in_node(stmt_idx, &mut arrows);
-            }
-        }
-        arrows
-    }
-
-    /// Check if constructor body needs `var _this = this;` capture
-    /// Returns true if the body contains arrow functions that capture `this`
-    fn constructor_needs_this_capture(&self, body_idx: NodeIndex) -> bool {
-        let arrow_indices = self.collect_arrow_functions_in_block(body_idx);
-
-        // Check if any arrow function captures `this`
-        for &arrow_idx in &arrow_indices {
-            if let Some(ref transforms) = self.transforms {
-                if let Some(crate::context::transform::TransformDirective::ES5ArrowFunction {
-                    captures_this,
-                    ..
-                }) = transforms.get(arrow_idx)
-                    && *captures_this
-                {
-                    return true;
-                }
-            } else {
-                // Fallback: directly check if arrow contains `this` reference
-                if contains_this_reference(self.arena, arrow_idx) {
-                    return true;
-                }
-            }
-        }
-
-        false
-    }
-
-    fn constructor_body_or_params_contain_new_target(
-        &self,
-        body_idx: NodeIndex,
-        params: &NodeList,
-    ) -> bool {
-        (body_idx.is_some() && contains_new_target_reference(self.arena, body_idx))
-            || params.nodes.iter().any(|&param_idx| {
-                self.arena
-                    .get(param_idx)
-                    .and_then(|param_node| self.arena.get_parameter(param_node))
-                    .is_some_and(|param| {
-                        param.initializer.is_some()
-                            && contains_new_target_reference(self.arena, param.initializer)
-                    })
-            })
-    }
-
-    fn class_constructor_new_target_capture_ir() -> IRNode {
-        IRNode::var_decl("_newTarget", Some(IRNode::Raw("this.constructor".into())))
-    }
-
-    fn insert_class_new_target_capture(&self, body: &mut Vec<IRNode>) {
-        let capture = Self::class_constructor_new_target_capture_ir();
-        if self.has_extends
-            && !self.extends_null
-            && let Some(super_capture_idx) = body
-                .iter()
-                .position(|node| self.is_generated_derived_super_capture(node))
-        {
-            body.insert(super_capture_idx + 1, capture);
-            return;
-        }
-
-        body.insert(0, capture);
-    }
-
-    fn is_generated_derived_super_capture(&self, node: &IRNode) -> bool {
-        let IRNode::VarDecl {
-            name,
-            initializer: Some(initializer),
-        } = node
-        else {
-            return false;
-        };
-        if name.as_ref() != "_this" {
-            return false;
-        }
-
-        matches!(
-            initializer.as_ref(),
-            IRNode::LogicalOr { left, right }
-                if matches!(right.as_ref(), IRNode::This { captured: false })
-                    && matches!(
-                        left.as_ref(),
-                        IRNode::CallExpr { callee, arguments }
-                            if arguments
-                                .first()
-                                .is_some_and(|arg| matches!(arg, IRNode::This { captured: false }))
-                                && matches!(
-                                    callee.as_ref(),
-                                    IRNode::PropertyAccess { object, property }
-                                        if property.as_ref() == "call"
-                                            && matches!(
-                                                object.as_ref(),
-                                                IRNode::Identifier(super_name)
-                                                    if super_name.as_ref() == self.super_name
-                                            )
-                                )
-                    )
-        )
-    }
-
-    fn instance_props_contain_new_target(&self, instance_props: &[NodeIndex]) -> bool {
-        instance_props.iter().any(|&prop_idx| {
-            self.arena
-                .get(prop_idx)
-                .and_then(|prop_node| self.arena.get_property_decl(prop_node))
-                .is_some_and(|prop| {
-                    prop.initializer.is_some()
-                        && contains_new_target_reference(self.arena, prop.initializer)
-                })
-        })
-    }
-
-    fn moved_instance_initializers_contain_new_target(&self, instance_props: &[NodeIndex]) -> bool {
-        self.instance_props_contain_new_target(instance_props)
-            || self.private_fields.iter().any(|field| {
-                !field.is_static
-                    && field.has_initializer
-                    && field.initializer.is_some()
-                    && contains_new_target_reference(self.arena, field.initializer)
-            })
-            || self.auto_accessors.iter().any(|accessor| {
-                !accessor.is_static
-                    && accessor.initializer.is_some_and(|initializer| {
-                        contains_new_target_reference(self.arena, initializer)
-                    })
-            })
-    }
-
-    /// Check if instance property initializers contain arrow functions that capture `this`.
-    /// Property initializers are moved into the constructor body by the ES5 transform.
-    fn instance_props_need_this_capture(&self, instance_props: &[NodeIndex]) -> bool {
-        for &prop_idx in instance_props {
-            let Some(prop_node) = self.arena.get(prop_idx) else {
-                continue;
-            };
-            let Some(prop_data) = self.arena.get_property_decl(prop_node) else {
-                continue;
-            };
-            if prop_data.initializer.is_none() {
-                continue;
-            }
-            // Check if the initializer contains arrow functions that capture `this`
-            let mut arrows = Vec::new();
-            self.collect_arrow_functions_in_node(prop_data.initializer, &mut arrows);
-            for &arrow_idx in &arrows {
-                if self
-                    .arena
-                    .get(arrow_idx)
-                    .and_then(|arrow_node| self.arena.get_function(arrow_node))
-                    .is_some_and(|arrow| arrow.is_async)
-                {
-                    return true;
-                }
-                if let Some(ref transforms) = self.transforms {
-                    if let Some(crate::context::transform::TransformDirective::ES5ArrowFunction {
-                        captures_this,
-                        ..
-                    }) = transforms.get(arrow_idx)
-                        && *captures_this
-                    {
-                        return true;
-                    }
-                } else if contains_this_reference(self.arena, arrow_idx) {
-                    return true;
-                }
-            }
-        }
-        false
-    }
-
-    fn property_initializer_has_equals(
-        &self,
-        member_node: &Node,
-        prop: &tsz_parser::parser::node::PropertyDeclData,
-    ) -> bool {
-        let Some(text) = self.source_text else {
-            return prop.initializer.is_some();
-        };
-        let Some(init_node) = self.arena.get(prop.initializer) else {
-            return false;
-        };
-        if prop.type_annotation.is_none() {
-            return true;
-        }
-
-        let start = member_node.pos as usize;
-        let end = (init_node.pos as usize).min(text.len());
-        if start >= end {
-            return false;
-        }
-        let segment = &text.as_bytes()[start..end];
-        let search_from = segment
-            .iter()
-            .rposition(|&byte| byte == b':')
-            .map_or(0, |idx| idx + 1);
-        segment[search_from..].contains(&b'=')
-    }
-
-    /// Recursively collect arrow function indices starting from a node
-    fn collect_arrow_functions_in_node(&self, idx: NodeIndex, arrows: &mut Vec<NodeIndex>) {
-        let Some(node) = self.arena.get(idx) else {
-            return;
-        };
-
-        if node.kind == syntax_kind_ext::ARROW_FUNCTION {
-            arrows.push(idx);
-        }
-
-        for child_idx in self.arena.get_children(idx) {
-            self.collect_arrow_functions_in_node(child_idx, arrows);
-        }
     }
 }
